refactor(orchestrator): move domains lib code to orchestrator
There was not really any reason for a separate library: `domains` to exist.
