+++ apps/compilers/project.json
@@ -2,6 +2,7 @@
   "$schema": "../../node_modules/nx/schemas/project-schema.json",
   "name": "compilers",
   "projectType": "application",
+  "implicitDependencies": ["examples"],
   "sourceRoot": "apps/compilers/src",
   "targets": {
     "build": {
