+++ apps/orchestrator/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod farem;
 pub mod graphql;
+pub mod users;
 
 use std::{env, sync::Arc};
 
@@ -9,6 +11,11 @@ use dotenv_codegen::dotenv;
 use edgedb_tokio::{Builder as DbBuilder, Client as DbClient};
 use figment::{providers::Env, Figment};
 use protobuf::generated::compilers::compiler_service_client::CompilerServiceClient;
+use rocket::{
+    async_trait,
+    request::{FromRequest, Outcome},
+    Request,
+};
 use surf::{Client as HttpClient, Config as HttpConfig, Url as HttpUrl};
 use tonic::transport::Channel;
 
@@ -62,4 +69,26 @@ impl AppConfig {
 pub async fn get_app_config() -> Result<AppConfig> {
     dotenv().ok();
     AppConfig::new().await
+}
+
+#[derive(Debug)]
+pub struct RequestData {
+    pub user_token: Option<String>,
+    pub jwt_secret: Vec<u8>,
+}
+
+pub struct Token(pub Option<String>);
+
+#[async_trait]
+impl<'r> FromRequest<'r> for Token {
+    type Error = String;
+
+    async fn from_request(request: &'r Request<'_>) -> Outcome<Self, Self::Error> {
+        let token_str = request
+            .headers()
+            .get_one("authorization")
+            .map(|f| f.to_string());
+        let token = Token(token_str);
+        Outcome::Success(token)
+    }
 }
