+++ apps/orchestrator/src/users/dto/queries/user_details.rs
