+++ apps/orchestrator/src/users/dto/queries/user_with_email.rs
