--- libs/domains/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "domains"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-anyhow = "1.0.65"
-async-graphql = { version = "4.0.14", features = ["uuid"] }
-async-trait = "0.1.57"
-auth = { path = "../auth" }
-config = { path = "../config" }
-derive-getters = "0.2.0"
-edgedb-derive = { git = "https://github.com/edgedb/edgedb-rust.git", branch = "master" }
-edgedb-errors = { git = "https://github.com/edgedb/edgedb-rust.git", branch = "master" }
-edgedb-protocol = { git = "https://github.com/edgedb/edgedb-rust.git", branch = "master" }
-edgedb-tokio = { git = "https://github.com/edgedb/edgedb-rust.git", branch = "master" }
-macros = { path = "../macros" }
-protobuf = { path = "../../libs/protobuf" }
-rocket = { version = "0.5.0-rc.2", features = ["json"] }
-serde = { version = "1.0.145", features = ["derive"] }
-serde_json = "1.0.85"
-strum = { version = "0.24.1", features = ["derive"] }
-surf = { version = "2.3.2", features = ["hyper-client"] }
-tonic = "0.8"
-uuid = "1.1.2"
