--- libs/domains/project.json
@@ -1,20 +1,0 @@
-{
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "projectType": "library",
-  "sourceRoot": "libs/domains/src",
-  "targets": {
-    "test": {
-      "executor": "@ignisda/nx-rust:nextest",
-      "options": {}
-    },
-    "lint": {
-      "executor": "@ignisda/nx-rust:clippy",
-      "options": {
-        "fix": false,
-        "failOnWarnings": true,
-        "noDeps": true
-      }
-    }
-  },
-  "tags": []
-}
